backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
google-generativeai==0.3.1